fix AuctionPessimistic, AuctionOptimistic implementations add null check to Notifier.sendOutdatedMessage
