+++ src/main/java/course/concurrency/exams/auction/AuctionOptimistic.java
@@ -1,69 +1,45 @@
 package course.concurrency.exams.auction;
 
-import java.util.concurrent.locks.StampedLock;
+import java.util.concurrent.atomic.AtomicReference;
 
 public class AuctionOptimistic implements Auction {
 
     private final Notifier notifier;
-    private final StampedLock stampedLock = new StampedLock();
 
     public AuctionOptimistic(Notifier notifier) {
         this.notifier = notifier;
     }
 
-    private volatile Bid latestBid;
+    private final AtomicReference<Bid> latestBid = new AtomicReference<>();
 
     /**
-     * Checks whether the new bid is the highest and updates the latest bid in this case
+     * Checks whether the new newBid is the highest and updates the latest newBid in this case
      *
-     * @param bid a new bid
-     * @return true if the latest bid was updated, false otherwise
+     * @param newBid a new newBid
+     * @return true if the latest newBid was updated, false otherwise
      */
-    public boolean propose(Bid bid) {
-        long stamp = stampedLock.tryOptimisticRead();
+    public boolean propose(Bid newBid) {
+        Bid currentBid = latestBid.get();
+        if (currentBid == null || newBid.getPrice() > currentBid.getPrice()) {
+            boolean updateSucceeded = latestBid.compareAndSet(currentBid, newBid);
 
-        if (latestBid == null || isNewBidHigher(bid)) {
-            if (stampedLock.validate(stamp)) { // checking if the latest bid was updated by another thread
-                updateLatestBid(bid);
+            while (!updateSucceeded) {
+                currentBid = latestBid.get();
+                if (newBid.getPrice() <= currentBid.getPrice()) // check if the new bid still beats the current one
+                    break;
+                updateSucceeded = latestBid.compareAndSet(currentBid, newBid);
+            }
+
+            if (updateSucceeded) {
+                notifier.sendOutdatedMessage(currentBid);
                 return true;
-            } else {
-                return updateLatestBidIfNeeded(bid);
             }
         }
         return false;
     }
 
-    public Bid getLatestBid() {
-        long stamp = stampedLock.tryOptimisticRead();
-        Bid lastAcceptedBid = latestBid;
-        if (!stampedLock.validate(stamp)) {
-            stamp = stampedLock.readLock();
-            lastAcceptedBid = latestBid;
-            stampedLock.unlockRead(stamp);
-        }
-        return lastAcceptedBid;
+    public Bid getLatestBid(){
+        return latestBid.get();
     }
 
-    private boolean isNewBidHigher(Bid bid) {
-        return bid.getPrice() > latestBid.getPrice();
-    }
-
-    private void updateLatestBid(Bid bid) {
-        long stamp = stampedLock.writeLock();
-        notifier.sendOutdatedMessage(latestBid);
-        latestBid = bid;
-        stampedLock.unlockWrite(stamp);
-    }
-
-    private boolean updateLatestBidIfNeeded(Bid bid) {
-        boolean isBidUpdated = false;
-        long stamp = stampedLock.writeLock();
-        if (isNewBidHigher(bid)) {
-            notifier.sendOutdatedMessage(latestBid);
-            latestBid = bid;
-            isBidUpdated = true;
-        }
-        stampedLock.unlockWrite(stamp);
-        return isBidUpdated;
-    }
 }
