+++ src/main/java/course/concurrency/exams/auction/AuctionPessimistic.java
@@ -1,5 +1,7 @@
 package course.concurrency.exams.auction;
 
+import java.util.concurrent.locks.StampedLock;
+
 public class AuctionPessimistic implements Auction {
 
     private final Notifier notifier;
@@ -8,18 +10,60 @@ public class AuctionPessimistic implements Auction {
         this.notifier = notifier;
     }
 
-    private Bid latestBid;
+    private final StampedLock stampedLock = new StampedLock();
+    private volatile Bid latestBid;
 
+    /**
+     * Checks whether the new bid is the highest and updates the latest bid in this case
+     *
+     * @param bid a new bid
+     * @return true if the latest bid was updated, false otherwise
+     */
     public boolean propose(Bid bid) {
-        if (latestBid == null || bid.getPrice() > latestBid.getPrice()) {
-            notifier.sendOutdatedMessage(latestBid);
-            latestBid = bid;
-            return true;
+        long stamp = stampedLock.tryOptimisticRead();
+
+        if (latestBid == null || isNewBidHigher(bid)) {
+            if (stampedLock.validate(stamp)) { // checking if the latest bid was updated by another thread
+                updateLatestBid(bid);
+                return true;
+            } else {
+                return updateLatestBidIfNeeded(bid);
+            }
         }
         return false;
     }
 
     public Bid getLatestBid() {
-        return latestBid;
+        long stamp = stampedLock.tryOptimisticRead();
+        Bid lastAcceptedBid = latestBid;
+        if (!stampedLock.validate(stamp)) {
+            stamp = stampedLock.readLock();
+            lastAcceptedBid = latestBid;
+            stampedLock.unlockRead(stamp);
+        }
+        return lastAcceptedBid;
+    }
+
+    private boolean isNewBidHigher(Bid bid) {
+        return bid.getPrice() > latestBid.getPrice();
+    }
+
+    private void updateLatestBid(Bid bid) {
+        long stamp = stampedLock.writeLock();
+        notifier.sendOutdatedMessage(latestBid);
+        latestBid = bid;
+        stampedLock.unlockWrite(stamp);
+    }
+
+    private boolean updateLatestBidIfNeeded(Bid bid) {
+        boolean isBidUpdated = false;
+        long stamp = stampedLock.writeLock();
+        if (isNewBidHigher(bid)) {
+            notifier.sendOutdatedMessage(latestBid);
+            latestBid = bid;
+            isBidUpdated = true;
+        }
+        stampedLock.unlockWrite(stamp);
+        return isBidUpdated;
     }
 }
